+++ docs/src/markdown/about/changelog.md
@@ -5,9 +5,7 @@ icon: lucide/scroll-text
 
 ## 8.11
 
--   **NEW**: Add a faster non-cached path for `oklch-cubic` that is optimized for when caching is disabled.
--   **NEW**: `algebra.clamp()` no longer accepts `None` for a minimum or maximum. Use `min()` or `max()` if no lower or
-    upper bound is provided.
+-   **ENHANCE**: Add a faster non-cached path for `oklch-cubic` that is optimized for when caching is disabled.
 -   **ENHANCE**: Improve HCT's conversion to XYZ by increasing the speed and quality of the results.
 -   **ENHANCE**: Improve RYB conversion to make it slightly faster and more accurate.